Fixed issues with map and searchbar
