+++ public/js/image-detail.js
