--- public/js/main.js
@@ -1,9 +1,0 @@
-const map = require('./components/map.js');
-const instructionSlides = require('./components/instructionSlides.js');
-const submitLocationTimestamp = require('./components/submitLocationTimestamp.js');
-
-(function () {
-  map.init();
-  instructionSlides.init();
-  submitLocationTimestamp.init();
-}) ();
